更新结构，按照ROS2标准写法重构，补充舵轮运动解算

+++ include/chassis_lib/steering_wheel.h
@@ -13,7 +13,7 @@ public:
     SteeringWheel(const int &can_id,const std::shared_ptr<Can> &can_handle=nullptr);
     ~SteeringWheel();
 
-    void sendCommand(const int &vel,const int &angle);
+    void sendCommand(const double &vel,const double &angle);
 };
