+++ src/steering_wheel.cpp
@@ -1,6 +1,7 @@
 #include "chassis_lib/steering_wheel.h"
 
-SteeringWheel::SteeringWheel(const int &can_id,const std::shared_ptr<Can> &can_handle) {
+SteeringWheel::SteeringWheel(const int &can_id,const std::shared_ptr<Can> &can_handle) 
+{
     this->can_id = can_id;
     if (can_handle == nullptr) {
         this->can_handle = std::make_shared<Can>("can0");
@@ -16,8 +17,8 @@ SteeringWheel::~SteeringWheel() {
     // Stop the PID controller
 }
 
-void SteeringWheel::sendCommand(const int &vel,const int &angle)
+void SteeringWheel::sendCommand(const double &vel,const double &angle)
 {
-    
+
 }
